Add files via upload
Outline: 

The change of all six images have been done by using the Adobe Photoshop program, which allowed to change and alter original photographs that have been saved from the web-browser as well as the media platform Pinterest. 
The main focus of changes has been in colour, contrasts and different effects and layers. As the authors Bomford (1995 ) and Gage (1999) state, colour plays a crucial role in the perception of the preferences whether something can be considered as aesthetics or not and add that many great works of art would less likely be as impressive in grey scale as they are in colour (in Albers, et al. 2020, p. 108). This can be seen in one of the edited images demonstrated above (Image 2) where a gradient, warm coloured circle in the background and a gradient warm coloured bar on the face of the model, gives the picture more live and context in terms of a different atmosphere.
As some colour combinations and their effects have been explained and analyse by Lupton and Phillips (2015), some theories are reflected and can be recognised in the edited images such as in Image 4 where a pastel and brighter form of two complementary contrast colours red/pink and blue/violet are in one composition but create a less biting cold/hot contrast than the actual colours on the spectrum would. This contrast is more effective on Image 3 where the background colour is predominantly containing a warm sunset orange and the outer frame of this is containing a violet filter. The hot-cold, dark-bright and complementary contrast in one photo gives a more aggressive impression, however, structures the foreground from the background. According to various authors, not only the object but also the colours human eye perceives as cold or hot for instance have a psychological effect on the viewer (Neale C., 2021). Using photoshop in order to change the brightness, sharpness and implementing effects on images is known to be improving engagement and quality of media (Bakhshi S. et al. 2015). 

Referring the design aspects of the images to Norman’s three levels of design, the most predominant factors that are reflected in the images are the visual and emotional aspects (Norman D., 2004). The emotional design relates to the emotion of the human mind, which eventually can be seen as something subjective, since different perspectives and behavioural levels can be triggered depending on the viewer. Emotional design is known to be connected with the emotional levels that are influenced by design. In image two and six the emotional aspect in the colour choices have been aimed to be triggered. 
The visual design aspect on the other hand, is referring to features, smaller details, and the quality of objects. The visual design aspects can be seen in images three four and five. The aspects of design are more predominant and appealing by different techniques that have been used in photoshop such as filters, that allow the add different textures, snipping that allow to create different ways to present and layers that allow to use different effects on the background and selected subjects. Image five for instance can be contextualised in both ways the visual as well as the emotional design aspect, since the composition may awake a personal, metaphoric meaning to the user and the quality of design may also be recognised as successful.